# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gift API documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get current user's profile",
                "description": "Get current user's profile",
                "operationId": "74bd0f9272b49f847e45d0f56cfc43a7",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "name"
                                        },
                                        "email": {
                                            "type": "email",
                                            "example": "email@mail.com"
                                        },
                                        "point": {
                                            "type": "number",
                                            "example": 10
                                        },
                                        "roles": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User Not Found."
                                        },
                                        "code": {
                                            "type": "number",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update current user's profile",
                "description": "Update current user's profile",
                "operationId": "50cc3de68a2a0e5304be9d710888c006",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "user1"
                                    },
                                    "email": {
                                        "type": "email",
                                        "format": "email",
                                        "example": "user1@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "name"
                                        },
                                        "email": {
                                            "type": "email",
                                            "example": "email@mail.com"
                                        },
                                        "point": {
                                            "type": "number",
                                            "example": 10
                                        },
                                        "roles": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User Not Found."
                                        },
                                        "code": {
                                            "type": "number",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong Credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to Update Current User's Profile."
                                        },
                                        "code": {
                                            "type": "number",
                                            "example": 500
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Something Wrong."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}